hex conversion module
